melhorando os mocks

+++ tests/CursoOnline.Dominio.UnitTests/Cursos/CursoDto.cs
@@ -4,7 +4,7 @@ namespace CursoOnline.Dominio.UnitTests.Cursos
     public class CursoDto
     {
         public string Nome { get; set; }
-        public int CargaHoraria { get; set; }
+        public double CargaHoraria { get; set; }
         public PublicoAlvo PublicoAlvo { get; set; }
         public double ValorCurso { get; set; }
         public string Descricao { get; set; }
